Rename alarm_serializer to alarm_parser
